feat: Connects OpenData from server
It turned out, the endpoint data needs cleaning. Each field was
apparently input as text, no enum or such. There aere a lot of
inconsistencies, which requires massive post-process to render
markers on map.

This commit is a proof of concept that web server gets data from
the endpoint and cleans it. After that, the web server returns
small JSON for JavaScript to render marker.